requirements.txt
openai
pydantic
python-docx
jinja2
